changed the index type for int64_t instead of uint64_t so I can use -1 as indicator of inexistent box. I will not be missing the later half of uint64 possible indexes, that's for sure

+++ sph_data_types.h
@@ -1,8 +1,8 @@
 #include <stdint.h>
 
 typedef struct size_t2{
-    size_t begin;
-    size_t end;
+    int64_t begin;
+    int64_t end;
 } size_t2;
 
 typedef struct double4 {
@@ -15,7 +15,7 @@ typedef struct double4 {
 typedef struct SPHparticle
 {
     double4 r,u,F;
-	size_t id,hash;
+	int64_t id,hash;
 	double nu,rho;
 } SPHparticle;
